do ignore sub rendering. fixes #2
There is no reliable way to figure out if we're rendering the current
main page, or if another renderer was instantiated during rendering the
page.

dokuwiki/dokuwiki#3267 addresses a similar issue in the handler. To
properly fix this, the page context would need to be passed to all
involved components during rendering and lots of code would need to be
checked to behave correctly. It's a mess.

This workaround checks the backtrace of our handler. When there are
multiple instances of p_render() we know that one render process
triggered another, nested one.